Code reorganised
