+++ ESPBASE.h
@@ -27,7 +27,9 @@
 #define ESPBASE_H
 
 #include <Arduino.h>
+
 #include <string.h>
+
 #ifdef ARDUINO_ESP32_DEV
     //ESP32 specific here
     #include <ESP32PWM.h>
@@ -51,15 +53,23 @@
   }
 #endif
 
-#include "global.h"
-#include "parameters.h"
 
-//#ifdef   ARDUINO_ESP8266_NODEMCU || ARDUINO_ESP8266_ESP01
+
+class ESPBASE {
+public:
+    bool WIFI_connected, CFG_saved;
+    uint8_t initialize();
+    void httpSetup();
+    void OTASetup();
+};
+
+#include "parameters.h"
+#include "global.h"
+#include "wifiTools.h"
 #include "NTP.h"
-//#endif
+
 
 // Include the HTML, STYLE and Script "Pages"
-
 #include "Page_Admin.h"
 #include "Page_Script.js.h"
 #include "Page_Style.css.h"
@@ -69,19 +79,7 @@
 #include "PAGE_NetworkConfiguration.h"
 
 
-#define LED_esp 2
-
-Ticker ticker;
-//os_timer_t myTimer;
-char tmpESP[100];
-
-class ESPBASE {
-public:
-    boolean WIFI_connected, CFG_saved;
-    uint8_t initialize();
-
-};
-
+//char tmpESP[100];
 
 uint8_t ESPBASE::initialize(){
 
@@ -102,9 +100,11 @@ uint8_t ESPBASE::initialize(){
   //**** Network Config load
   CFG_saved = ReadConfig();
 
-  //  Connect to WiFi acess point or start as Acess point
+  //  Connect to WiFi access point or start as Access point
   if (CFG_saved) { //if config saved use saved values
+
       // Connect the ESP8266 to local WIFI network in Station mode
+      // using SSID and password saved in parameters (config object)
       Serial.println("Booting");
       //printConfig();
       WiFi.mode(WIFI_OFF);
@@ -134,112 +134,117 @@ uint8_t ESPBASE::initialize(){
 
     // DEFAULT CONFIG
     Serial.println("Setting AP mode default parameters");
-    #ifdef ARDUINO_ESP32_DEV
-      config.ssid = "ESP32-" + String(getChipId(),HEX);       // SSID of access point
-    #elif ARDUINO_ESP8266_ESP01 || ARDUINO_ESP8266_NODEMCU
-      config.ssid = "ESP8266-" + String(getChipId(),HEX);       // SSID of access point
-    #endif
-    config.password = "" ;   // password of access point
-    config.dhcp = true;
-    config.IP[0] = 192; config.IP[1] = 168; config.IP[2] = 1; config.IP[3] = 100;
-    config.Netmask[0] = 255; config.Netmask[1] = 255; config.Netmask[2] = 255; config.Netmask[3] = 0;
-    config.Gateway[0] = 192; config.Gateway[1] = 168; config.Gateway[2] = 1; config.Gateway[3] = 254;
-    config.ntpServerName = "0.ch.pool.ntp.org"; //"0.ch.pool.ntp.org"; // to be adjusted to PT ntp.ist.utl.pt
-    config.Update_Time_Via_NTP_Every =  5;
-    config.timeZone = 1;
-    config.isDayLightSaving = true;
-    config.DeviceName = "Not Named";
-    config.OTApwd = "";
+
+    //load config with default values
+    configLoadDefaults(getChipId());
+
     WiFi.mode(WIFI_AP);
     WiFi.softAP(config.ssid.c_str());
     Serial.print("Wifi ip:");Serial.println(WiFi.softAPIP());
    }
 
+    //  Http Setup
 
-    // Start HTTP Server for configuration
-    server.on ( "/", []() {
-      Serial.println("admin.html");
-      server.send_P ( 200, "text/html", PAGE_AdminMainPage);  // const char top of page
-    }  );
-    server.on ( "/favicon.ico",   []() {
-      Serial.println("favicon.ico");
-      server.send( 200, "text/html", "" );
-    }  );
-    // Network config
-    server.on ( "/config.html", send_network_configuration_html );
-    // Info Page
-    server.on ( "/info.html", []() {
-      Serial.println("info.html");
-      server.send_P ( 200, "text/html", PAGE_Information );
-    }  );
-    server.on ( "/ntp.html", send_NTP_configuration_html  );
+    httpSetup();
 
-    //server.on ( "/appl.html", send_application_configuration_html  );
-    server.on ( "/general.html", send_general_html  );
-    //  server.on ( "/example.html", []() { server.send_P ( 200, "text/html", PAGE_EXAMPLE );  } );
-    server.on ( "/style.css", []() {
-      Serial.println("style.css");
-      server.send_P ( 200, "text/plain", PAGE_Style_css );
-    } );
-    server.on ( "/microajax.js", []() {
-      Serial.println("microajax.js");
-      server.send_P ( 200, "text/plain", PAGE_microajax_js );
-    } );
-    server.on ( "/admin/values", send_network_configuration_values_html );
-    server.on ( "/admin/connectionstate", send_connection_state_values_html );
-    server.on ( "/admin/infovalues", send_information_values_html );
-    server.on ( "/admin/ntpvalues", send_NTP_configuration_values_html );
-    //server.on ( "/admin/applvalues", send_application_configuration_values_html );
-    server.on ( "/admin/generalvalues", send_general_configuration_values_html);
-    server.on ( "/admin/devicename",     send_devicename_value_html);
-    server.onNotFound ( []() {
-      Serial.println("Page Not Found");
-      server.send ( 400, "text/html", "Page not Found" );
-    }  );
-    server.begin();
-    Serial.println( "HTTP server started" );
     // ***********  OTA SETUP
-
-    //ArduinoOTA.setHostname(host);
-    ArduinoOTA.onStart([]() { // what to do before OTA download insert code here
-        Serial.println("Start");
-      });
-    ArduinoOTA.onProgress([](unsigned int progress, unsigned int total) {
-      Serial.printf("Progress: %u%%\r", (progress / (total / 100)));
-    });
-    ArduinoOTA.onEnd([]() { // do a fancy thing with our board led at end
-        for (int i=0;i<30;i++){
-          analogWrite(LED_esp,(i*100) % 1001);
-          delay(50);
-        }
-        digitalWrite(LED_esp,HIGH); // Switch OFF ESP LED to save energy
-        Serial.println("\nEnd");
-        ESP.restart();
-      });
-
-    ArduinoOTA.onError([](ota_error_t error) {
-        Serial.printf("Error[%u]: ", error);
-        if (error == OTA_AUTH_ERROR) Serial.println("Auth Failed");
-        else if (error == OTA_BEGIN_ERROR) Serial.println("Begin Failed");
-        else if (error == OTA_CONNECT_ERROR) Serial.println("Connect Failed");
-        else if (error == OTA_RECEIVE_ERROR) Serial.println("Receive Failed");
-        else if (error == OTA_END_ERROR) Serial.println("End Failed");
-        ESP.restart();
-      });
-
-     /* setup the OTA server */
-    ArduinoOTA.setPassword(config.OTApwd.c_str());    // ********    to be implemented as a parameter via Browser
-    ArduinoOTA.begin();
-    Serial.println("Ready");
+    OTASetup();
 
 
+    // start internal time update ISR
     #ifdef ARDUINO_ESP32_DEV
       tkSecond.attach(1000000, ISRsecondTick);
     #elif ARDUINO_ESP8266_ESP01 || ARDUINO_ESP8266_NODEMCU
-      // start internal time update ISR
       tkSecond.attach(1, ISRsecondTick);
     #endif
+
+    Serial.println("Ready");
+
 }
+
+
+void ESPBASE::httpSetup(){
+  // Start HTTP Server for configuration
+  server.on ( "/", []() {
+    Serial.println("admin.html");
+    server.send_P ( 200, "text/html", PAGE_AdminMainPage);  // const char top of page
+  }  );
+  server.on ( "/favicon.ico",   []() {
+    Serial.println("favicon.ico");
+    server.send( 200, "text/html", "" );
+  }  );
+  // Network config
+  server.on ( "/config.html", send_network_configuration_html );
+  // Info Page
+  server.on ( "/info.html", []() {
+    Serial.println("info.html");
+    server.send_P ( 200, "text/html", PAGE_Information );
+  }  );
+  server.on ( "/ntp.html", send_NTP_configuration_html  );
+
+  //server.on ( "/appl.html", send_application_configuration_html  );
+  server.on ( "/general.html", send_general_html  );
+  //  server.on ( "/example.html", []() { server.send_P ( 200, "text/html", PAGE_EXAMPLE );  } );
+  server.on ( "/style.css", []() {
+    Serial.println("style.css");
+    server.send_P ( 200, "text/plain", PAGE_Style_css );
+  } );
+  server.on ( "/microajax.js", []() {
+    Serial.println("microajax.js");
+    server.send_P ( 200, "text/plain", PAGE_microajax_js );
+  } );
+  server.on ( "/admin/values", send_network_configuration_values_html );
+  server.on ( "/admin/connectionstate", send_connection_state_values_html );
+  server.on ( "/admin/infovalues", send_information_values_html );
+  server.on ( "/admin/ntpvalues", send_NTP_configuration_values_html );
+  //server.on ( "/admin/applvalues", send_application_configuration_values_html );
+  server.on ( "/admin/generalvalues", send_general_configuration_values_html);
+  server.on ( "/admin/devicename",     send_devicename_value_html);
+  server.onNotFound ( []() {
+    Serial.println("Page Not Found");
+    server.send ( 400, "text/html", "Page not Found" );
+  }  );
+  server.begin();
+  Serial.println( "HTTP server started" );
+
+  return;
+}
+
+void ESPBASE::OTASetup(){
+
+      //ArduinoOTA.setHostname(host);
+      ArduinoOTA.onStart([]() { // what to do before OTA download insert code here
+          Serial.println("Start");
+        });
+      ArduinoOTA.onProgress([](unsigned int progress, unsigned int total) {
+        Serial.printf("Progress: %u%%\r", (progress / (total / 100)));
+      });
+      ArduinoOTA.onEnd([]() { // do a fancy thing with our board led at end
+          for (int i=0;i<30;i++){
+            analogWrite(LED_esp,(i*100) % 1001);
+            delay(50);
+          }
+          digitalWrite(LED_esp,HIGH); // Switch OFF ESP LED to save energy
+          Serial.println("\nEnd");
+          ESP.restart();
+        });
+
+      ArduinoOTA.onError([](ota_error_t error) {
+          Serial.printf("Error[%u]: ", error);
+          if (error == OTA_AUTH_ERROR) Serial.println("Auth Failed");
+          else if (error == OTA_BEGIN_ERROR) Serial.println("Begin Failed");
+          else if (error == OTA_CONNECT_ERROR) Serial.println("Connect Failed");
+          else if (error == OTA_RECEIVE_ERROR) Serial.println("Receive Failed");
+          else if (error == OTA_END_ERROR) Serial.println("End Failed");
+          ESP.restart();
+        });
+
+       /* setup the OTA server */
+      ArduinoOTA.setPassword(config.OTApwd.c_str());    // ********    to be implemented as a parameter via Browser
+      ArduinoOTA.begin();
+
+}
+
 
 
 #endif
