+++ Page_NetworkConfiguration.h
@@ -73,6 +73,18 @@ Please Wait....Configuring and Restarting.
 //  SEND HTML PAGE OR IF A FORM SUMBITTED VALUES, PROCESS THESE VALUES
 //
 
+// Check the Values is between 0-255
+boolean checkRange(String Value){
+  if (Value.toInt() < 0 || Value.toInt() > 255)
+  {
+    return false;
+  }
+  else
+  {
+    return true;
+  }
+}
+
 void send_network_configuration_html()
 {
